ALEA-844 Remove MISRA and improve compiler diagnostics https://gitlab.com/alea-2020/command-data-handling/obc2-firmware/-/merge_requests/150
# Overview
- Disabled the MISRA checker since it had so many false positives and made the code significantly more difficult to read
- Removed all references to OBC_MISRA_OFF/ON and obc_misra.h :tada:
- Set compiler to emit warnings as errors
- Set compiler to emit remarks
- Fixed existing remarks (and disabled some non-relevant ones)
  - Codebase currently compiles for both platforms with no remarks. Future MRs will be rejected if they introduce any new remarks (unless there are special circumstances).

# Testing
HIL tests passing

+++ obc-firmware/main/app/orcasat/system/settings/obc_settings.c
@@ -308,7 +308,7 @@ static int32_t setting_exists(const char* setting_name) {
         return -1;
     }
 
-    int32_t i = 0;
+    uint32_t i = 0;
     for (i = 0; i < num_settings; i++) {
         if (strcmp(settings[i].name, setting_name) == 0) {
             return i;
@@ -325,7 +325,7 @@ static int32_t setting_exists(const char* setting_name) {
  */
 static bool increment_revision_count(void) {
     bool ok = false;
-    if ((revision_count_index != -1) && (revision_count_index < num_settings)) {
+    if ((revision_count_index != -1) && ((uint32_t)revision_count_index < num_settings)) {
         setting_t* revision_setting = &settings[revision_count_index];
         revision_setting->setting_value++;
